Anmol Home,Login Sign up/sr-comsoft/backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
mysql-connector-python==9.1.0
bcrypt==4.2.1
PyJWT==2.10.1
python-dotenv==1.0.1